fix: wrap parameter editor with Accordion
