+++ services/orchest-webserver/client/src/styles/main.scss
@@ -313,7 +313,6 @@ p > i.material-icons.float-left {
 .CodeMirror {
   font-size: $codeSize;
   border-radius: 4px;
-  padding: 12px 0;
   height: auto;
 }
